+++ tests/conftest.py
@@ -156,9 +156,11 @@ async def central_unit_pydevccu_full(pydevccu_full: pydevccu.Server) -> CentralU
 
     central = await get_pydev_ccu_central_unit_full(port=const.CCU_PORT)
 
-    unsubscribe_homematic_callback = central.event_bus.subscribe(event_type=HomematicEvent, handler=homematic_callback)
+    unsubscribe_homematic_callback = central.event_bus.subscribe(
+        event_type=HomematicEvent, event_key=None, handler=homematic_callback
+    )
     unsubscribe_backend_system_callback = central.event_bus.subscribe(
-        event_type=BackendSystemEventData, handler=backend_system_callback
+        event_type=BackendSystemEventData, event_key=None, handler=backend_system_callback
     )
 
     try: